Update location IDs to Int64, fix crashing bug when deleting currently viewed location

+++ sweather-2/Content View/ContentView.swift
@@ -93,7 +93,7 @@ struct ContentView_Previews: PreviewProvider {
 extension WWLocation {
     init(savedLocation: SavedLocation) {
         self = WWLocation(
-            id: Int(savedLocation.id),
+            id: Int64(savedLocation.id),
             name: savedLocation.name!,
             region: savedLocation.region!,
             state: savedLocation.state!,
